spi-ft232h: register nrc80211 spi device for GW16146
Add spi_board_info struct to driver_info so that we can define SPI
device details and register the SPI device. Also add fields for
GPIO reset.

Configure this for GW16146 which uses an FT232H as a SPI master
connected to an NRC7292 with an out-of-tree driver that binds as
'nrc80211'.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>